add new class for healthcheck
