+++ ruby-scanner-scaffolding.gemspec
@@ -4,7 +4,7 @@ Gem::Specification.new do |s|
   s.date        = '2019-04-15'
   s.summary     = "Gem providing camunda worker for integrated, ruby based, scanner"
   s.authors     = ["Iteratec GmbH"]
-  s.files       = ["lib/ruby-scanner-scaffolding.rb"]
+  s.files       = ["lib/ruby-scanner-scaffolding.rb", "lib/ruby-scanner-scaffolding/healthcheck.rb"]
   s.homepage    =
       'https://github.com/secureCodeBox/ruby-scanner-scaffolding'
 end
